requirement.txt
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
bson==0.5.8
colorama==0.4.3
dataclasses==0.6
Django==2.2.13
django-betterforms==1.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-mapbox-location-field==1.5.0
django-taggit==1.3.0
djongo==1.3.2
dnspython==1.16.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
